requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
python-dotenv
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
streamlit>=1.30.0
